pyproject.toml
[tool.poetry]
name = "cryppiweb3"
version = "0.1.0"
description = ""
authors = ["hhhscvx <hhhscvx@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.20.1"
python-dotenv = "^1.0.1"
fake-useragent = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
